LayoutTests/fast/dom/Range/getClientRects-expected.txt

Test 1
PASS rects.length is 5
PASS rects[0].left.toFixed(2) is "8.00"
PASS rects[0].top.toFixed(2) is "8.00"
PASS rects[0].width.toFixed(2) is "400.00"
PASS rects[0].height.toFixed(2) is "160.00"
PASS rects[1].left.toFixed(2) is "8.00"
PASS rects[1].top.toFixed(2) is "19.00"
PASS rects[1].width.toFixed(2) is "386.58"
PASS rects[1].height.toFixed(2) is "18.00"
PASS rects[2].left.toFixed(2) is "8.00"
PASS rects[2].top.toFixed(2) is "59.00"
PASS rects[2].width.toFixed(2) is "359.91"
PASS rects[2].height.toFixed(2) is "18.00"
PASS rects[3].left.toFixed(2) is "8.00"
PASS rects[3].top.toFixed(2) is "99.00"
PASS rects[3].width.toFixed(2) is "381.25"
PASS rects[3].height.toFixed(2) is "18.00"
PASS rects[4].left.toFixed(2) is "8.00"
PASS rects[4].top.toFixed(2) is "139.00"
PASS rects[4].width.toFixed(2) is "368.36"
PASS rects[4].height.toFixed(2) is "18.00"
Test 2
PASS rects.length is 4
PASS rects[0].left.toFixed(2) is "8.00"
PASS rects[0].top.toFixed(2) is "215.00"
PASS rects[0].width.toFixed(2) is "386.58"
PASS rects[0].height.toFixed(2) is "18.00"
PASS rects[1].left.toFixed(2) is "8.00"
PASS rects[1].top.toFixed(2) is "255.00"
PASS rects[1].width.toFixed(2) is "359.91"
PASS rects[1].height.toFixed(2) is "18.00"
PASS rects[2].left.toFixed(2) is "8.00"
PASS rects[2].top.toFixed(2) is "295.00"
PASS rects[2].width.toFixed(2) is "381.25"
PASS rects[2].height.toFixed(2) is "18.00"
PASS rects[3].left.toFixed(2) is "8.00"
PASS rects[3].top.toFixed(2) is "335.00"
PASS rects[3].width.toFixed(2) is "368.36"
PASS rects[3].height.toFixed(2) is "18.00"
Test 3
PASS rects.length is 4
PASS rects[0].left.toFixed(2) is "81.31"
PASS rects[0].top.toFixed(2) is "411.00"
PASS rects[0].width.toFixed(2) is "313.27"
PASS rects[0].height.toFixed(2) is "18.00"
PASS rects[1].left.toFixed(2) is "8.00"
PASS rects[1].top.toFixed(2) is "451.00"
PASS rects[1].width.toFixed(2) is "359.91"
PASS rects[1].height.toFixed(2) is "18.00"
PASS rects[2].left.toFixed(2) is "8.00"
PASS rects[2].top.toFixed(2) is "491.00"
PASS rects[2].width.toFixed(2) is "381.25"
PASS rects[2].height.toFixed(2) is "18.00"
PASS rects[3].left.toFixed(2) is "8.00"
PASS rects[3].top.toFixed(2) is "531.00"
PASS rects[3].width.toFixed(2) is "146.66"
PASS rects[3].height.toFixed(2) is "18.00"
Test 4
PASS rects.length is 7
PASS rects[0].left.toFixed(2) is "8.00"
PASS rects[0].top.toFixed(2) is "596.00"
PASS rects[0].width.toFixed(2) is "400.00"
PASS rects[0].height.toFixed(2) is "278.00"
PASS rects[1].left.toFixed(2) is "8.00"
PASS rects[1].top.toFixed(2) is "607.00"
PASS rects[1].width.toFixed(2) is "386.58"
PASS rects[1].height.toFixed(2) is "18.00"
PASS rects[2].left.toFixed(2) is "8.00"
PASS rects[2].top.toFixed(2) is "725.00"
PASS rects[2].width.toFixed(2) is "268.39"
PASS rects[2].height.toFixed(2) is "18.00"
PASS rects[3].left.toFixed(2) is "352.39"
PASS rects[3].top.toFixed(2) is "725.00"
PASS rects[3].width.toFixed(2) is "44.88"
PASS rects[3].height.toFixed(2) is "18.00"
PASS rects[4].left.toFixed(2) is "8.00"
PASS rects[4].top.toFixed(2) is "765.00"
PASS rects[4].width.toFixed(2) is "348.38"
PASS rects[4].height.toFixed(2) is "18.00"
PASS rects[5].left.toFixed(2) is "8.00"
PASS rects[5].top.toFixed(2) is "805.00"
PASS rects[5].width.toFixed(2) is "375.90"
PASS rects[5].height.toFixed(2) is "18.00"
PASS rects[6].left.toFixed(2) is "8.00"
PASS rects[6].top.toFixed(2) is "845.00"
PASS rects[6].width.toFixed(2) is "67.98"
PASS rects[6].height.toFixed(2) is "18.00"
Test 5
PASS rects.length is 5
PASS rects[0].left.toFixed(2) is "81.00"
PASS rects[0].top.toFixed(2) is "921.00"
PASS rects[0].width.toFixed(2) is "313.58"
PASS rects[0].height.toFixed(2) is "18.00"
PASS rects[1].left.toFixed(2) is "8.00"
PASS rects[1].top.toFixed(2) is "1039.00"
PASS rects[1].width.toFixed(2) is "268.39"
PASS rects[1].height.toFixed(2) is "18.00"
PASS rects[2].left.toFixed(2) is "276.39"
PASS rects[2].top.toFixed(2) is "950.00"
PASS rects[2].width.toFixed(2) is "76.00"
PASS rects[2].height.toFixed(2) is "103.00"
PASS rects[3].left.toFixed(2) is "352.39"
PASS rects[3].top.toFixed(2) is "1039.00"
PASS rects[3].width.toFixed(2) is "44.88"
PASS rects[3].height.toFixed(2) is "18.00"
PASS rects[4].left.toFixed(2) is "8.00"
PASS rects[4].top.toFixed(2) is "1079.00"
PASS rects[4].width.toFixed(2) is "291.00"
PASS rects[4].height.toFixed(2) is "18.00"
Test 6
PASS rects.length is 8
PASS rects[0].left.toFixed(2) is "8.00"
PASS rects[0].top.toFixed(2) is "1224.00"
PASS rects[0].width.toFixed(2) is "400.00"
PASS rects[0].height.toFixed(2) is "240.00"
PASS rects[1].left.toFixed(2) is "8.00"
PASS rects[1].top.toFixed(2) is "1235.00"
PASS rects[1].width.toFixed(2) is "386.58"
PASS rects[1].height.toFixed(2) is "18.00"
PASS rects[2].left.toFixed(2) is "8.00"
PASS rects[2].top.toFixed(2) is "1275.00"
PASS rects[2].width.toFixed(2) is "16.00"
PASS rects[2].height.toFixed(2) is "18.00"
PASS rects[3].left.toFixed(2) is "8.00"
PASS rects[3].top.toFixed(2) is "1315.00"
PASS rects[3].width.toFixed(2) is "58.22"
PASS rects[3].height.toFixed(2) is "18.00"
PASS rects[4].left.toFixed(2) is "106.22"
PASS rects[4].top.toFixed(2) is "1315.00"
PASS rects[4].width.toFixed(2) is "45.32"
PASS rects[4].height.toFixed(2) is "18.00"
PASS rects[5].left.toFixed(2) is "8.00"
PASS rects[5].top.toFixed(2) is "1355.00"
PASS rects[5].width.toFixed(2) is "396.34"
PASS rects[5].height.toFixed(2) is "18.00"
PASS rects[6].left.toFixed(2) is "8.00"
PASS rects[6].top.toFixed(2) is "1395.00"
PASS rects[6].width.toFixed(2) is "359.93"
PASS rects[6].height.toFixed(2) is "18.00"
PASS rects[7].left.toFixed(2) is "8.00"
PASS rects[7].top.toFixed(2) is "1435.00"
PASS rects[7].width.toFixed(2) is "221.71"
PASS rects[7].height.toFixed(2) is "18.00"
Test 7
PASS rects.length is 8
PASS rects[0].left.toFixed(2) is "8.00"
PASS rects[0].top.toFixed(2) is "1511.00"
PASS rects[0].width.toFixed(2) is "386.58"
PASS rects[0].height.toFixed(2) is "18.00"
PASS rects[1].left.toFixed(2) is "8.00"
PASS rects[1].top.toFixed(2) is "1551.00"
PASS rects[1].width.toFixed(2) is "16.00"
PASS rects[1].height.toFixed(2) is "18.00"
PASS rects[2].left.toFixed(2) is "8.00"
PASS rects[2].top.toFixed(2) is "1580.00"
PASS rects[2].width.toFixed(2) is "400.00"
PASS rects[2].height.toFixed(2) is "40.00"
PASS rects[3].left.toFixed(2) is "8.00"
PASS rects[3].top.toFixed(2) is "1591.00"
PASS rects[3].width.toFixed(2) is "58.22"
PASS rects[3].height.toFixed(2) is "18.00"
PASS rects[4].left.toFixed(2) is "106.22"
PASS rects[4].top.toFixed(2) is "1591.00"
PASS rects[4].width.toFixed(2) is "45.32"
PASS rects[4].height.toFixed(2) is "18.00"
PASS rects[5].left.toFixed(2) is "8.00"
PASS rects[5].top.toFixed(2) is "1631.00"
PASS rects[5].width.toFixed(2) is "396.34"
PASS rects[5].height.toFixed(2) is "18.00"
PASS rects[6].left.toFixed(2) is "8.00"
PASS rects[6].top.toFixed(2) is "1671.00"
PASS rects[6].width.toFixed(2) is "359.93"
PASS rects[6].height.toFixed(2) is "18.00"
PASS rects[7].left.toFixed(2) is "8.00"
PASS rects[7].top.toFixed(2) is "1711.00"
PASS rects[7].width.toFixed(2) is "221.71"
PASS rects[7].height.toFixed(2) is "18.00"
Test 8
PASS rects.length is 4
PASS rects[0].left.toFixed(2) is "168.51"
PASS rects[0].top.toFixed(2) is "1758.44"
PASS rects[0].width.toFixed(2) is "262.28"
PASS rects[0].height.toFixed(2) is "307.71"
PASS rects[1].left.toFixed(2) is "137.87"
PASS rects[1].top.toFixed(2) is "1784.15"
PASS rects[1].width.toFixed(2) is "245.14"
PASS rects[1].height.toFixed(2) is "287.28"
PASS rects[2].left.toFixed(2) is "107.23"
PASS rects[2].top.toFixed(2) is "1809.86"
PASS rects[2].width.toFixed(2) is "258.85"
PASS rects[2].height.toFixed(2) is "303.62"
PASS rects[3].left.toFixed(2) is "76.59"
PASS rects[3].top.toFixed(2) is "1835.57"
PASS rects[3].width.toFixed(2) is "250.57"
PASS rects[3].height.toFixed(2) is "293.75"
Test 9
PASS rects.length is 0
Test 9b
PASS rects.length is 1
PASS rects[0].left.toFixed(2) is "8.00"
PASS rects[0].top.toFixed(2) is "1983.00"
PASS rects[0].width.toFixed(2) is "0.00"
PASS rects[0].height.toFixed(2) is "18.00"
Test 10
PASS rects.length is 5
PASS rects[0].left.toFixed(2) is "19.00"
PASS rects[0].top.toFixed(2) is "2168.00"
PASS rects[0].width.toFixed(2) is "18.00"
PASS rects[0].height.toFixed(2) is "375.00"
PASS rects[1].left.toFixed(2) is "59.00"
PASS rects[1].top.toFixed(2) is "2168.00"
PASS rects[1].width.toFixed(2) is "18.00"
PASS rects[1].height.toFixed(2) is "367.03"
PASS rects[2].left.toFixed(2) is "99.00"
PASS rects[2].top.toFixed(2) is "2168.00"
PASS rects[2].width.toFixed(2) is "18.00"
PASS rects[2].height.toFixed(2) is "348.38"
PASS rects[3].left.toFixed(2) is "139.00"
PASS rects[3].top.toFixed(2) is "2168.00"
PASS rects[3].width.toFixed(2) is "18.00"
PASS rects[3].height.toFixed(2) is "375.90"
PASS rects[4].left.toFixed(2) is "179.00"
PASS rects[4].top.toFixed(2) is "2168.00"
PASS rects[4].width.toFixed(2) is "18.00"
PASS rects[4].height.toFixed(2) is "67.98"
Test 11
PASS rects.length is 2
PASS rects[0].left.toFixed(2) is "8.00"
PASS rects[0].top.toFixed(2) is "2604.00"
PASS rects[0].width.toFixed(2) is "400.00"
PASS rects[0].height.toFixed(2) is "40.00"
PASS rects[1].left.toFixed(2) is "8.00"
PASS rects[1].top.toFixed(2) is "2615.00"
PASS rects[1].width.toFixed(2) is "181.75"
PASS rects[1].height.toFixed(2) is "18.00"
Test 12
PASS rects.length is 1
PASS rects[0].left.toFixed(2) is "8.00"
PASS rects[0].top.toFixed(2) is "2840.00"
PASS rects[0].width.toFixed(2) is "400.00"
PASS rects[0].height.toFixed(2) is "160.00"
Test 13
PASS rects.length is 4
PASS rects[0].left.toFixed(2) is "8.00"
PASS rects[0].top.toFixed(2) is "3047.00"
PASS rects[0].width.toFixed(2) is "386.58"
PASS rects[0].height.toFixed(2) is "18.00"
PASS rects[1].left.toFixed(2) is "8.00"
PASS rects[1].top.toFixed(2) is "3087.00"
PASS rects[1].width.toFixed(2) is "359.91"
PASS rects[1].height.toFixed(2) is "18.00"
PASS rects[2].left.toFixed(2) is "8.00"
PASS rects[2].top.toFixed(2) is "3127.00"
PASS rects[2].width.toFixed(2) is "381.25"
PASS rects[2].height.toFixed(2) is "18.00"
PASS rects[3].left.toFixed(2) is "8.00"
PASS rects[3].top.toFixed(2) is "3167.00"
PASS rects[3].width.toFixed(2) is "368.36"
PASS rects[3].height.toFixed(2) is "18.00"
Test 14
PASS rects.length is 0
PASS successfullyParsed is true

TEST COMPLETE
